+++ src/main/java/com/florian_ligneul/canbus/change_highlighter/service/reader/ACanBusMessageListener.java
@@ -1,27 +1,17 @@
 package com.florian_ligneul.canbus.change_highlighter.service.reader;
 
 import com.fazecast.jSerialComm.SerialPort;
-import com.fazecast.jSerialComm.SerialPortMessageListener;
+import com.fazecast.jSerialComm.SerialPortDataListener;
 
 /**
- * Event based message reception implementation of the {@link SerialPortMessageListener} with LF message delimiter.
+ * Event based message reception implementation of the {@link SerialPortDataListener}
+ * Callback will be triggered whenever some amount of data has actually been read from the serial port.
  */
-public abstract class ACanBusMessageListener implements SerialPortMessageListener {
-    private static byte[] LF_DELIMITER = new byte[]{(byte) 0x0A};
+public abstract class ACanBusMessageListener implements SerialPortDataListener {
 
     @Override
     public int getListeningEvents() {
         return SerialPort.LISTENING_EVENT_DATA_RECEIVED;
-    }
-
-    @Override
-    public byte[] getMessageDelimiter() {
-        return LF_DELIMITER;
-    }
-
-    @Override
-    public boolean delimiterIndicatesEndOfMessage() {
-        return true;
     }
 }